LayoutTests/fast/events/clientXY-in-zoom-and-scroll.html

<script src="../../resources/js-test-pre.js"></script>
<style>
    #pusher {
        width: 1000px;
        height: 1000px;
        outline: 1px solid black;
    }
    #subpixel-test {
        width: 4.5px;
    }
</style>
<div id="console"></div>
<div id="testArea">
    <div id="pusher">This box is here to create scrollbars.</div>
    <div id="subpixel-test"></div>
</div>
<script>
    var r = document.getElementById('subpixel-test').getBoundingClientRect();
    var hasSubpixelSupport = r.right - r.left == 4.5;

    var event;

    async function sendClick()
    {
        if (window.eventSender) {
            await eventSender.asyncMouseMoveTo(100, 100);
            await eventSender.asyncMouseDown();
            await eventSender.asyncMouseUp();
        }
    }

    function zoomPageIn()
    {
        if (window.eventSender) {
            eventSender.zoomPageIn();
        }
    }

    function zoomPageOut()
    {
        if (window.eventSender) {
            eventSender.zoomPageOut();
        }
    }

    function scrollPage(x, y)
    {
        window.scrollTo(x, y);
    }

    if (window.testRunner) {
        testRunner.dumpAsText();
        testRunner.waitUntilDone();
    }

    // Default.
    function base(e)
    {
        event = e;
        debug("Base");
        shouldBe("event.clientX", "100");
        shouldBe("event.clientY", "100");
        shouldBe("event.pageX", "100");
        shouldBe("event.pageY", "100");
    }
    
    // Just zoomed.
    function justZoomed(e)
    {
        event = e;
        debug("\nJust zoomed");
        shouldBe("event.clientX", "83");
        shouldBe("event.clientY", "83");
        shouldBe("event.pageX", "83");
        shouldBe("event.pageY", "83");
    }

    // Just scrolled.
    function justScrolled(e)
    {
        event = e;
        debug("\nJust scrolled");
        shouldBe("event.clientX", "100");
        shouldBe("event.clientY", "100");
        shouldBe("event.pageX", "150");
        shouldBe("event.pageY", "150");
    }

    // Zoomed and scrolled.
    function zoomedAndScrolled(e)
    {
        event = e;
        debug("\nZoomed and scrolled");
        shouldBe("event.clientX", hasSubpixelSupport ? "83" : "84");
        shouldBe("event.clientY", hasSubpixelSupport ? "83" : "84");
        shouldBe("event.pageX", "133");
        shouldBe("event.pageY", "133");
    }

    onload = async () => {
        if (!window.testRunner)
            return;

        testRunner.dumpAsText();
        testRunner.waitUntilDone();

        window.addEventListener("click", base, false);
        await sendClick();
        window.removeEventListener("click", base, false);

        window.addEventListener("click", justZoomed, false);
        zoomPageIn();
        await sendClick();
        zoomPageOut();
        window.removeEventListener("click", justZoomed, false);

        window.addEventListener("click", justScrolled, false);
        scrollPage(50, 50);
        await sendClick();
        scrollPage(0, 0);
        window.removeEventListener("click", justScrolled, false);

        window.addEventListener("click", zoomedAndScrolled, false);
        zoomPageIn();
        scrollPage(50, 50);
        await sendClick();
        zoomPageOut();
        scrollPage(0, 0);
        window.removeEventListener("click", zoomedAndScrolled, false);

        var area = document.getElementById('testArea');
        area.parentNode.removeChild(area);

        testRunner.notifyDone();
    }
</script>
<script src="../../resources/js-test-post.js"></script>
